LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/interfaces/HTMLElement/HTMLTrackElement/srclang-expected.txt

CONSOLE MESSAGE: The language contains a null character and is not a valid BCP 47 language tag.
CONSOLE MESSAGE: The language ' foo
' is not a valid BCP 47 language tag.
CONSOLE MESSAGE: The language ' foo
' is not a valid BCP 47 language tag.
CONSOLE MESSAGE: The language contains a null character and is not a valid BCP 47 language tag.

PASS HTMLTrackElement.srclang missing value
PASS HTMLTrackElement.srclang empty string content attribute
PASS HTMLTrackElement.srclang empty string IDL attribute
PASS HTMLTrackElement.srclang lowercase content attribute
PASS HTMLTrackElement.srclang uppercase content attribute
PASS HTMLTrackElement.srclang \u0000 content attribute
PASS HTMLTrackElement.srclang lowercase IDL attribute
PASS HTMLTrackElement.srclang uppercase IDL attribute
PASS HTMLTrackElement.srclang whitespace in content attribute
PASS HTMLTrackElement.srclang whitespace in IDL attribute
PASS HTMLTrackElement.srclang \u0000 in IDL attribute
